backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-decouple==3.8

django-filter==23.3
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
requests==2.31.0
africastalking==2.0

python-multipart==0.0.6
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2

dj-database-url
# django-mptt
django-mptt==0.15.0

redis==5.0.1






python-dotenv
